requirements.txt
Cython
numpy>=1.16.0
scipy>=1.2.0
scikit-learn
keras>=2.3.0
gensim>=3.2.0
pandas
pulp
PyStemmer
flask